Added error route for unmatched routes. Renamed health check to be health check.
